Move header to another file
