+++ src/app/clusters/water-heater-management-server/BUILD.gn
@@ -16,6 +16,8 @@ import("//build_overrides/chip.gni")
 
 source_set("water-heater-management-server") {
   sources = [
+    "Delegate.cpp",
+    "Delegate.h",
     "WaterHeaterManagementCluster.cpp",
     "WaterHeaterManagementCluster.h",
     "WaterHeaterManagementTestEventTriggerHandler.h",
