+++ src/app/clusters/water-heater-management-server/Delegate.cpp
@@ -1,0 +1,51 @@
+/*
+ *
+ *    Copyright (c) 2024 Project CHIP Authors
+ *    All rights reserved.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+#include <app/clusters/water-heater-management-server/Delegate.h>
+#include <app/clusters/water-heater-management-server/WaterHeaterManagementCluster.h>
+
+namespace chip {
+namespace app {
+namespace Clusters {
+namespace WaterHeaterManagement {
+
+/***************************************************************************
+ *
+ * Delegate event forwarding
+ *
+ ***************************************************************************/
+
+CHIP_ERROR Delegate::GenerateBoostStartedEvent(uint32_t durationSecs, Optional<bool> oneShot, Optional<bool> emergencyBoost,
+                                               Optional<int16_t> temporarySetpoint, Optional<Percent> targetPercentage,
+                                               Optional<Percent> targetReheat)
+{
+    VerifyOrReturnError(mCluster != nullptr, CHIP_ERROR_INCORRECT_STATE);
+    return mCluster->GenerateBoostStartedEvent(durationSecs, oneShot, emergencyBoost, temporarySetpoint, targetPercentage,
+                                               targetReheat);
+}
+
+CHIP_ERROR Delegate::GenerateBoostEndedEvent()
+{
+    VerifyOrReturnError(mCluster != nullptr, CHIP_ERROR_INCORRECT_STATE);
+    return mCluster->GenerateBoostEndedEvent();
+}
+
+} // namespace WaterHeaterManagement
+} // namespace Clusters
+} // namespace app
+} // namespace chip
